+++ kornia/color/__init__.py
@@ -16,6 +16,7 @@ from .yuv import RgbToYuv, YuvToRgb, rgb_to_yuv, yuv_to_rgb
 from .xyz import RgbToXyz, XyzToRgb, rgb_to_xyz, xyz_to_rgb
 from .luv import RgbToLuv, LuvToRgb, rgb_to_luv, luv_to_rgb
 from .normalize import Normalize, normalize, Denormalize, denormalize
+from .histogram import histogram, histogram2d
 from .core import add_weighted, AddWeighted
 from .adjust import (
     AdjustBrightness, AdjustContrast, AdjustGamma, AdjustHue, AdjustSaturation,
@@ -44,6 +45,8 @@ __all__ = [
     "xyz_to_rgb",
     "normalize",
     "denormalize",
+    "histogram",
+    "histogram2d",
     "adjust_brightness",
     "adjust_contrast",
     "adjust_gamma",
