+++ test/color/test_histogram.py
